Adding support for protobufs and libprotobuf-mutator
